researched and recreated code
